download: Move utils function into test file
